+++ src/check_jsonschema/loaders/schema/__init__.py
@@ -142,14 +142,22 @@ class MetaSchemaLoader(SchemaLoaderBase):
         return validator(validator.META_SCHEMA)
 
 
-def schema_loader_from_args(args) -> SchemaLoaderBase:
-    if args.schemafile is not None:
-        return SchemaLoader(
-            args.schemafile,
-            args.cache_filename,
-            args.no_cache,
-        )
-    elif args.builtin_schema:
-        return BuiltinSchemaLoader(args.builtin_schema)
-    else:
+def schema_loader_from_args(
+    schemafile: str | None,
+    cache_filename: str | None,
+    no_cache: bool,
+    builtin_schema: str | None,
+    check_metaschema: bool,
+) -> SchemaLoaderBase:
+    if check_metaschema:
         return MetaSchemaLoader()
+    elif builtin_schema is not None:
+        return BuiltinSchemaLoader(builtin_schema)
+    if schemafile is not None:
+        return SchemaLoader(
+            schemafile,
+            cache_filename,
+            no_cache,
+        )
+    else:
+        raise NotImplementedError("no valid schema option provided")