+++ tests/acceptance/conftest.py
@@ -1,51 +1,29 @@
 import pytest
+from click.testing import CliRunner
 
-from check_jsonschema import main
-
-
-class _CLIResult:
-    def __init__(self):
-        self.err = None
-        self.exit_code = 0
-        self.stdout = ""
-        self.stderr = ""
-
-    def __str__(self):
-        return f"""\
-<CLIResult>
-err: {self.err}
-exit_code: {self.exit_code}
-<stdout>
-{self.stdout}
-</stdout>
-<stderr>
-{self.stderr}
-</stderr>
-</CLIResult>
-"""
+from check_jsonschema import main as cli_main
 
 
 @pytest.fixture
-def cli_runner(capsys):
-    def func(args, *, expect_ok=True):
-        res = _CLIResult()
+def cli_runner():
+    return CliRunner(mix_stderr=False)
 
-        try:
-            main(args)
-        except SystemExit as err:
-            res.exit_code = err.code
-            res.err = err
-        except Exception as err:
-            res.exit_code = 1
-            res.err = err
 
-        captured = capsys.readouterr()
-        res.stdout = captured.out
-        res.stderr = captured.err
+@pytest.fixture
+def run_line(cli_runner):
+    def func(cli_args, *args, **kwargs):
+        assert cli_args[0] == "check-jsonschema"
+        if "catch_exceptions" not in kwargs:
+            kwargs["catch_exceptions"] = False
+        return cli_runner.invoke(cli_main, cli_args[1:], *args, **kwargs)
 
-        if expect_ok:
-            assert res.exit_code == 0, str(res)
+    return func
 
-        return res
+
+@pytest.fixture
+def run_line_simple(run_line):
+    def func(cli_args, *args, **kwargs):
+        res = run_line(["check-jsonschema"] + cli_args, *args, **kwargs)
+        assert res.exit_code == 0
 
     return func