+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 <!-- vendor-insert-here -->
 - Update vendored schemas (2022-05-02)
+- Use `click` to implement CLI parsing. This provides several internal features
+  as well as shell completion support.
 
 ## 0.14.3